requirements.txt
pillow
pyside6
pyinstaller